+++ src/LDL/Validators/ArrayTypeValidator.php
@@ -3,16 +3,13 @@
 namespace LDL\Validators;
 
 use LDL\Validators\Exception\TypeMismatchException;
+use LDL\Validators\Traits\ValidatorHasConfigInterfaceTrait;
 use LDL\Validators\Traits\ValidatorValidateTrait;
 
-class ArrayTypeValidator implements ValidatorInterface, NegatedValidatorInterface
+class ArrayTypeValidator implements ValidatorInterface, NegatedValidatorInterface, ValidatorHasConfigInterface
 {
     use ValidatorValidateTrait;
-
-    /**
-     * @var Config\BasicValidatorConfig
-     */
-    private $config;
+    use ValidatorHasConfigInterfaceTrait;
 
     public function __construct(
         bool $negated = false,
@@ -20,7 +17,7 @@ class ArrayTypeValidator implements ValidatorInterface, NegatedValidatorInterfac
         string $description=null
     )
     {
-        $this->config = new Config\BasicValidatorConfig($negated, $dumpable, $description);
+        $this->_tConfig = new Config\BasicValidatorConfig($negated, $dumpable, $description);
     }
 
     public function assertTrue($value): void
@@ -73,13 +70,5 @@ class ArrayTypeValidator implements ValidatorInterface, NegatedValidatorInterfac
          * @var Config\ValidatorConfigInterface $config
          */
         return new self($config->isNegated(), $config->isDumpable());
-    }
-
-    /**
-     * @return Config\BasicValidatorConfig
-     */
-    public function getConfig(): Config\BasicValidatorConfig
-    {
-        return $this->config;
     }
 }